Rename folder to gencmdxml
